rename appConstants to userConstants
